the slow decay of happy thoughts GLbuffer uses named buffer because thats a thing. means less annoying binding.
brrrrrrrr

completely lost track of the changes made. last looked at this months ago.

scene has been split into the various systems using funky diamond virtual protected inheritance thingies.

material got replaced by specular material. it now uses c++20 wierd concept thingies for the maps. idk if thats rly necessary or a good use but hey what ever. I'm likely gonna replace material again with a super material which will take uniform buffers and allocate them a base to bind to.

The vbos are now components added to each entity and when the entity fulfills an archetype, the vao in the material is initiated. This is done with entt events but could probably be done better by setting up a view or group.

the vbo object needs to be changed somewhat to allow either for copying or for sharing of resource. I can do this by storing a pointer to the first and next shared resource as a form of reference counting.

There is currently no error handling for incomplete mesh structures.

I moved the entt macros to a their own header file.

model loading has been removed. the old code is in notes.txt

+++ TentilliumEngine/Source/Rendering/Resources/ShaderProgram.cpp
@@ -56,20 +56,18 @@ ShaderProgram::ShaderProgram(std::string filepath)
 {
 	std::string fp(filepath);
 
-	int p1 = fp.find_last_of("\\/");
-	int p2 = fp.find_last_of(".");
+	size_t p1 = fp.find_last_of("\\/");
+	size_t p2 = fp.find_last_of(".");
 
 	std::string dirPath = fp.substr(0, p1 == std::string::npos ? 0 : p1);
-	std::string dirName = fp.substr(
-		p1 == std::string::npos ? 0 : p1 + 1,
-		(p2 == std::string::npos ? fp.length() : (p2 - 1)) - p1);
+	std::string dirName = fp.substr(p1 == std::string::npos ? 0 : p1 + 1, (p2 == std::string::npos ? fp.length() : (p2 - 1)) - p1);
 
 	ShaderLink shaders;
 	for (const auto& file : std::filesystem::directory_iterator(dirPath)) {
 		std::string fp = file.path().string();
 
-		int p1 = fp.find_last_of("\\/");
-		int p2 = fp.find_last_of(".");
+		size_t p1 = fp.find_last_of("\\/");
+		size_t p2 = fp.find_last_of(".");
 
 		if (p1 == std::string::npos) p1 = 0; else p1 += 1;
 		if (p2 == std::string::npos) p2 = fp.length();
